+++ examples/state_multiple/src/main.rs
@@ -13,6 +13,21 @@ struct Counter {
 
 type SharedCounter = Rc<RefCell<Counter>>;
 
+async fn get_counters(
+    State((a_counter, b_counter)): State<(SharedCounter, SharedCounter)>,
+) -> impl IntoResponse {
+    #[derive(serde::Serialize)]
+    struct Response {
+        a: Counter,
+        b: Counter,
+    }
+
+    picoserve::response::Json(Response {
+        a: a_counter.borrow().clone(),
+        b: b_counter.borrow().clone(),
+    })
+}
+
 async fn get_counter(State(state): State<SharedCounter>) -> impl IntoResponse {
     picoserve::response::Json(state.borrow().clone())
 }
@@ -27,16 +42,31 @@ async fn set_counter(value: i32, State(state): State<SharedCounter>) -> impl Int
     Redirect::to(".")
 }
 
+fn make_app<State>(
+    counter: SharedCounter,
+) -> picoserve::Router<impl picoserve::routing::PathRouter<State>, State> {
+    picoserve::Router::new()
+        .route("/", get(get_counter))
+        .route("/increment", get(increment_counter))
+        .route(("/set", parse_path_segment()), get(set_counter))
+        .with_state(counter)
+}
+
 #[tokio::main(flavor = "current_thread")]
 async fn main() -> anyhow::Result<()> {
     let port = 8000;
 
+    let a_counter = Rc::new(RefCell::new(Counter { counter: 0 }));
+    let b_counter = Rc::new(RefCell::new(Counter { counter: 0 }));
+
     let app = std::rc::Rc::new(
         picoserve::Router::new()
-            .route("/", get(get_counter))
-            .route("/increment", get(increment_counter))
-            .route(("/set", parse_path_segment()), get(set_counter))
-            .with_state(Rc::new(RefCell::new(Counter { counter: 0 }))),
+            .route("/", get(get_counters))
+            .route("/a", get(|| async { Redirect::to("/a/") }))
+            .route("/b", get(|| async { Redirect::to("/b/") }))
+            .nest("/a", make_app(a_counter.clone()))
+            .nest("/b", make_app(b_counter.clone()))
+            .with_state((a_counter, b_counter)),
     );
 
     let config = picoserve::Config::new(picoserve::Timeouts {